Backend/requirements.txt
Django
django-allauth
django-cors-headers
djangorestframework
filelock
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-generativeai
googleapis-common-protos
oauthlib
requests
requests-oauthlib
cryptography
jwt
